refactor: split off monodeploy-versions (#237)
* refactor: prepare monodeploy-versions

* refactor: fix imports

* refactor: clear up dead code
